- Warning fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33320 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/animesh/controllers/ragdoll/ragdoll.cpp
@@ -217,7 +217,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(Ragdoll)
     if (state != RAGDOLL_STATE_DYNAMIC)
       return 0;
 
-    return bones.GetSize ();
+    return (uint)bones.GetSize ();
   }
 
   BoneID RagdollAnimNode::GetBone (csChainStateType state, uint index) const
